Convert global.argv0 to a D slice
This is more of a proof of concept for future conversion.

Here's a few principles used to do the conversion from `T*` to `T[]`:
- If it's a string, the character at `s[$]` must be `'\0'`.
  It allows code to freely go back and forth between `char*` and `char[]`.
- Move the implementation to extern(D) and slices, and provide C++ wrapper.
- Stay on topic: those changes only focus on one parameter / module / function,
  because it's very easy to end up with a giant change that doesn't work.
- If `'\0'` termination is required (e.g. calling a C function),
  changing from `char*` to `char[]` should be delayed.

The main offender in this commit is `Filename.exists`, which is hard to port,
given it needs `\0`-terminated string. This will be addressed in its own commit.