Make the Outbuffer / message function tuple-variadic and slice-aware
This turn the ubiquitous first layer of `message`, `warning`,
`deprecation[Supplemental]`, `error` from C-style variadic into D-style variadic,
while moving the previous implementation as `cmessage`, `cwarning`, etc...

The intent is to make it easy to print a slice using those functions.
At the moment, when one which to print a slice, one can not rely on `\0` termination
of the slice, so the following code:
```
error(loc, "Identifier not found: %s", id.toChars());
```
has to be turned into:
```
error(loc, "Identifier not found: %.*s", cast(int)id.toString().length, id.toString().ptr);
```

This is obviously not convenient and a major burden on the call site.
Worse, some `toString` might perform some computation (e.g. strlen at the moment)
and calling it might not be trivial, which usually leads to storing the slice as an extra variable.

As the goal of our conversion from `char*` to `char[]` is to improve the code quality,
a better solution is obviously needed.
With this patchset, the previous call to `error` becomes:
```
error(loc, "Identifier not found: %.*s", id.toString());
```
As a result, only the format string has to be altered.