Deprecate `extern(C++, identifier)` in favor of `extern(C++, "str")`
This features tries to import C++ semantics into D, by giving namespaces a scope.
However, when it comes to interfacing with C++, we only care about namespaces
in the context of mangling: they shouldn't be used for code organization.

Namespaces don't work so well for D: for example, namespaces in C++ can be spread across multiple files,
which is something D can't and won't do because it conflicts with its very core, modules.
It also leads to issues where identifiers used for namespaces
(which are usually outside of the programmer's control) end up conflicting with user code.

Instead, users should use `extern(C++, "string")`,
which makes D code follows D semantic while letting users link with a symbol in a namespace.